+++ parser/config.go
@@ -139,6 +139,7 @@ var SupportedFunctions = map[string]int{
 	"sigma":     1,
 	"isperfect": 1,
 	"digitsum":   1,
+	"primorial": 1,
 	"digitcount": 1, // Collatz stopping time
 	"prod":      -1,
 	"min":       -1, // variadic